Revert "WiP (this breaks stuff) -> Trying to avoid using os.execute"
This reverts commit 01a91fd2a2441e68beb3267419596a0de1ca8d93.
